+++ guillotina/db/cache/__init__.py
@@ -1,1 +1,2 @@
 from . import dummy  # noqa
+from . import base  # noqa