+++ guillotina/db/interfaces.py
@@ -21,6 +21,7 @@ class ITransaction(Interface):
     user = Attribute('')
     status = Attribute('')
     storage = Attribute('')
+    manager = Attribute('')
     _cache = Attribute('')
 
     async def add_after_commit_hook(hook, *real_args, args=[], kws=None, **kwargs):
@@ -109,6 +110,38 @@ class ITransactionManager(Interface):
     async def abort(*, txn: typing.Optional[ITransaction]=None):
         '''
         abort txn
+        '''
+
+class ITransactionCache(Interface):
+
+    async def clear():  # type: ignore
+        '''
+        clear cache
+        '''
+
+    async def get(oid=None, container=None, id=None, variant=None):
+        '''
+        get cached object
+        '''
+
+    async def set(value, oid=None, container=None, id=None, variant=None):
+        '''
+        set cached data
+        '''
+
+    async def delete(key):
+        '''
+        delete cache key
+        '''
+
+    async def delete_all(keys):
+        '''
+        delete list of keys
+        '''
+
+    async def close():  # type: ignore
+        '''
+        close the cache
         '''
 
 
@@ -271,39 +304,6 @@ class ITransactionStrategy(Interface):
 
 class IDBTransactionStrategy(ITransactionStrategy):
     pass
-
-
-class IStorageCache(Interface):
-
-    async def clear():  # type: ignore
-        '''
-        clear cache
-        '''
-
-    async def get(oid=None, container=None, id=None, variant=None):
-        '''
-        get cached object
-        '''
-
-    async def set(value, oid=None, container=None, id=None, variant=None):
-        '''
-        set cached data
-        '''
-
-    async def delete(key):
-        '''
-        delete cache key
-        '''
-
-    async def delete_all(keys):
-        '''
-        delete list of keys
-        '''
-
-    async def close():  # type: ignore
-        '''
-        close the cache
-        '''
 
 
 class IDatabaseManager(Interface):