+++ guillotina/interfaces/__init__.py
@@ -1,7 +1,9 @@
 # zope.interface convenience imports
 from .async_util import IAsyncJobPool  # noqa
 from .async_util import IAsyncUtility  # noqa
+from .async_util import ICacheUtility  # noqa
 from .async_util import IQueueUtility  # noqa
+from .async_util import IPubSubUtility  # noqa
 from .behaviors import IAsyncBehavior  # noqa
 from .behaviors import IBehavior  # noqa
 from .behaviors import IBehaviorAdapterFactory  # noqa
@@ -31,10 +33,12 @@ from .content import IResourceFactory  # noqa
 from .content import IStaticDirectory  # noqa
 from .content import IStaticFile  # noqa
 from .content import ITraversable  # noqa
+from .events import IAfterAsyncUtilityLoadedEvent  # noqa
 from .events import IApplicationCleanupEvent  # noqa
 from .events import IApplicationConfiguredEvent  # noqa
 from .events import IApplicationEvent  # noqa
 from .events import IApplicationInitializedEvent  # noqa
+from .events import IBeforeAsyncUtilityLoadedEvent  # noqa
 from .events import IBeforeObjectAddedEvent  # noqa
 from .events import IBeforeObjectModifiedEvent  # noqa
 from .events import IBeforeObjectMovedEvent  # noqa